Make DV use BA guides
